pyproject.toml
[tool.poetry]
name = "vone-drl"
version = "0.1.0"
description = "Solving Virtual Optical Network Embedding using Deep Reinforcement Learning"
authors = ["Michael Doherty, Yitao Zhang"]
readme = "README.md"
packages = [{include = "vone_drl"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Gymnasium = "^0.26.3"
numpy = "^1.23.4"
stable-baselines3 = "^1.6.2"
networkx = "^2.8.7"
Cython = "^0.29.32"
pymongo = "^4.3.2"
pillow = "^9.3.0"
tqdm = "^4.64.1"
scipy = "^1.9.3"
joblib = "^1.2.0"
pip = "^22.2"
sklearn = "^0.0"
geneticalgorithm = "^1.0.2"
progress = "^1.6"
cffi = "^1.15.1"
wandb = "^0.13.4"
sb3-contrib = "^1.6.2"
nvidia-cuda-runtime-cu11 = {version = "^2022.4.25", markers = "sys_platform != 'darwin'"}
tensorflow = "^2.10.0"
ray = {version = "^2.1.0", markers = "sys_platform != 'win64'"}
scikit-learn = "^1.1.3"
sympy = "^1.11.1"
black = "^22.12.0"
pytest = "^7.2.0"

[[tool.poetry.source]]
name = "nvidia"
url = "https://pypi.ngc.nvidia.com"
default = false
secondary = true

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
default = true

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/foo.py  # exclude a file named foo.py in the root of the project (in addition to the defaults)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
